+++ index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import ReactDOM from 'react-dom';
 
-import Main from './components/Main1';
+import Main from './components/Main';
 
 class App extends Component {
 	render() {
